+++ src/ua/leonidius/taylor/algorithms/TaylorAlgorithm.java
@@ -6,6 +6,6 @@ import java.math.BigDecimal;
 
 public interface TaylorAlgorithm {
 
-    BigDecimal compute(MathFunction function, BigDecimal argument, int iterationsNum);
+    BigDecimal compute(MathFunction function, double argument, int iterationsNum);
 
 }
